add lua and micropython

+++ pkgs/bl808-rootfs.nix
@@ -1,4 +1,4 @@
-{ stdenv, fakeroot, squashfsTools, pkgsCross, writeReferencesToFile, linkFarmFromDrvs, asDir ? false }:
+{ stdenv, fakeroot, squashfsTools, pkgsCross, writeReferencesToFile, linkFarmFromDrvs, asDir ? false, python3, pkg-config }:
 let
   pkgsTarget = pkgsCross.riscv64.pkgsMusl;
   pkgsTargetStatic = pkgsCross.riscv64.pkgsStatic;
@@ -21,9 +21,33 @@ stdenv.mkDerivation rec {
       #"--without-tzdata"
     ];
   });
+  micropython = (pkgsTarget.micropython.override { python3 = python3; }).overrideAttrs (old: {
+    # We need a compiler for the host to build mpy-cross. The build complains that pkg-config is missing so let's add that, as well.
+    nativeBuildInputs = (old.nativeBuildInputs or []) ++ [ stdenv.cc pkg-config ];
+
+    buildPhase = ''
+      runHook preBuild
+      make -C mpy-cross -j$NIX_BUILD_CORES
+
+      # - use cross-compiler here (but not for mpy-cross)
+      # - don't set Q= to the cross-compiler prefix because that's also used to call Python
+      # - remove -Werror from CWARN
+      make -C ports/unix CC=${pkgsTarget.stdenv.cc.targetPrefix}gcc CROSS_COMPILE=${pkgsTarget.stdenv.cc.targetPrefix} CWARN=-Wall -j$NIX_BUILD_CORES
+      runHook postBuild
+    '';
+  });
+  lua = pkgsTarget.lua.override {
+    # no readline because that pulls in ncurses, which is too big
+    readline = null;
+    postConfigure = ''
+      makeFlagsArray+=(PLAT=linux MYLIBS=-ldl -j$NIX_BUILD_CORES)
+      echo "#undef LUA_USE_READLINE" >>./src/luaconf.h
+      substituteInPlace src/Makefile --replace "-lreadline" ""
+    '';
+  };
 
   #refsDrv = linkFarmFromDrvs "refs" [ busybox tcl ];
-  refsDrv = linkFarmFromDrvs "refs" [ busybox ];
+  refsDrv = linkFarmFromDrvs "refs" [ busybox micropython lua ];
   refs = writeReferencesToFile refsDrv;
 
   buildImage = ''
@@ -38,12 +62,18 @@ stdenv.mkDerivation rec {
     while read x ; do
       cp -r $x nix/store/
     done <$refs
-    rm -rf nix/store/*/{lib/*.a,lib/*.o,man/}
+    rm -rf nix/store/*/{lib/*.a,lib/*.o,lib/pkgconfig,man/,share/terminfo}
 
     mkdir bin
     ln -s bin sbin
     for x in nix/store/*/bin/* ; do
-      ln -s "/$x" "bin/''${x##*/}"
+      if [ ! -e "bin/''${x##*/}" ] ; then
+        ln -s "/$x" "bin/''${x##*/}"
+      else
+        echo "WARN: File already exists: bin/''${x##*/}"
+        echo "  points to $(realpath "bin/''${x##*/}")"
+        echo "  ignoring  /$x"
+      fi
     done
     ln -s ${busybox}/linuxrc .
     ln -s ${busybox}/default.script .
@@ -84,4 +114,6 @@ stdenv.mkDerivation rec {
       cp squashfs_test.img $out
     fi
   '';
+
+  dontFixup = true;
 }
